pyproject.toml
[project]
name = "data-parser"
version = "0.1.0"
description = "HTML to Markdown converter for GCS files"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=2.10.0",
    "beautifulsoup4>=4.12.0",
    "markitdown>=0.0.1",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0",
    "psycopg[binary,pool]>=3.2.9",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "testcontainers>=3.7.0",
    "requests>=2.31.0",
    "ruff>=0.12.11",
    "mypy>=1.17.1",
]
